refactoring protocol layers
refactoring protocol layers, adding extra layer
